장바구니 수정 전

+++ src/pages/admin/adminbook.js
@@ -33,45 +33,49 @@ function Adminbook() {
     }, [id]);
 
     if(!book) {
-        return <div>Book Not Found!</div>;
+        return <h2>Book Loading...</h2>
     }
 
     const deleteBook = async (id) => {
-        try {
-          await axios.delete(`/api/books/${id}`);
-          alert('삭제되었습니다.');
-          navigate("/admin/adminbooks")
-        } catch (error) {
-          console.error("삭제에 실패했습니다:", error);
-          alert('삭제에 실패했습니다.');
+        const isConfirmed = window.confirm("책을 삭제 하시겠습니까?");
+
+        if(isConfirmed) {
+            try {
+            await axios.delete(`/api/books/${id}`);
+            alert('삭제되었습니다.');
+            navigate("/admin/adminbooks")
+            } catch (error) {
+            console.error("삭제에 실패했습니다:", error);
+            alert('삭제에 실패했습니다.');
+            window.location.reload();
+            }            
         }
+        else {
+            window.location.reload();
+        }
+
     }
 
     const deletecomment = async (id) => {
-        try {
+        const isConfirmed = window.confirm("댓글을 삭제 하시겠습니까?");
+
+        if(isConfirmed) {
+            try {
             await axios.delete(`/api/comments/${id}`);
             alert('댓글이 삭제되었습니다.');
+            window.location.reload();
           } catch (error) {
             console.error("삭제에 실패했습니다:", error);
             alert('댓글 삭제에 실패했습니다.');
+            window.location.reload();
           }
+        }
+        else {
+            window.location.reload();
+        }
+        
     }
 
-    const saveComment = async (id) => {
-        try {
-            await axios.patch(`/api/comments/${id}`, {
-                "id" : editingCommentId,
-                "bookId": bookid,
-                "content": editingContent
-            });
-            alert('댓글이 수정되었습니다.');
-        } catch (error) {
-            console.error("현재 id: "+id + "현재 commentid : " +editingCommentId)
-            alert('댓글 수정에 실패했습니다.');
-        }
-    };
-    
-    
     return (
         <html>
             <head>
@@ -82,7 +86,7 @@ function Adminbook() {
                 <div className="jumbotron"> 
                     <div className="container">
                         <h1 className="display-3" align="center">관리자 도서 정보</h1>
-                        <h5 className="display-5" align="center">Book Details</h5>
+                        <h5 className="display-5" align="center">Admin Book Details</h5>
                     </div>
                 </div>
 
@@ -119,8 +123,8 @@ function Adminbook() {
                                     <div className="card-body">
                                         {editingCommentId === comment.id ? (
                                             <>
-                                                <input value={editingContent} onChange={(e) => setEditingContent(e.target.value)} className="form-control"/>
-                                                <button className="btn btn-success" onClick={() => saveComment(comment.id)}>저장</button>
+                                                {/* <input value={editingContent} onChange={(e) => setEditingContent(e.target.value)} className="form-control"/> */}
+                                                {/* <button className="btn btn-success" onClick={() => saveComment(comment.id)}>저장</button> */}
                                             </>
                                         ) : (
                                             <>
@@ -143,58 +147,10 @@ function Adminbook() {
 
                             <br />
 
-                            {/* <b>댓글 작성</b>
-                            <form onSubmit={createcomment} className="form-horizontal">
-                                <div className="d-flex">
-                                    <div className="col-sm-8">
-                                        <input name="content" value={newcomment} onChange={(e) => setNewcomment(e.target.value)} className="form-control" placeholder="댓글을 입력하세요"/>
-                                    </div>
-                                    <div className="col-sm-2">  
-                                        <button type="submit" className="btn btn-primary">등록</button>      
-                                    </div>
-                                </div>                                         
-                            </form> */}
-
-
                             <br />
-                            <Link to={`/books/book/update/${book.id}`} className="btn btn-success" role="button">책 수정 &raquo;</Link>
+                            <Link to={`/books/book/update/${book.id}`} className="btn btn-success" role="button">책 수정 &raquo;</Link> &nbsp;
                             <Link to={`/books/book/delete/${book.id}`} className="btn btn-danger"  onClick={() => deleteBook(book.id)} role="button">책 삭제 &raquo;</Link>
 
-                            <br />
-                            <br />
-
-                            {/* <div>
-                                <label htmlFor="purchaseQuantity">장바구니 수량 선택 : </label>
-                                <select 
-                                    id="purchaseQuantity"
-                                    value={cartQuantity} 
-                                    onChange={e => setCartQuantity(e.target.value)} 
-                                    className="form-control" 
-                                    style={{width: 'auto', display: 'inline-block', margin: '0 10px'}}
-                                >
-                                    {[...Array(10).keys()].map(n => (
-                                        <option key={n + 1} value={n + 1}>{n + 1}</option>
-                                    ))}
-                                </select>
-                                <button className="btn btn-success" onClick={cart}>장바구니에 추가 &raquo;</button>
-                            </div>
-                            
-                            <div>
-                                <label htmlFor="purchaseQuantity">구매 수량 선택 : </label>
-                                <select 
-                                    id="purchaseQuantity"
-                                    value={purchaseQuantity} 
-                                    onChange={e => setPurchaseQuantity(e.target.value)} 
-                                    className="form-control" 
-                                    style={{width: 'auto', display: 'inline-block', margin: '0 10px'}}
-                                >
-                                    {[...Array(10).keys()].map(n => (
-                                        <option key={n + 1} value={n + 1}>{n + 1}</option>
-                                    ))}
-                                </select>
-                                <button className="btn btn-danger" onClick={purchase}>바로 구매하기 &raquo;</button>
-                            </div> */}
-                            
                             <br />
                             <br />
