+++ src/pages/book/book.js
@@ -16,6 +16,22 @@ function Book() {
     const [editingContent, setEditingContent] = useState("");
     const [bookid, setBookid] = useState(1);
 
+    const increaseQuantity = () => {
+      setCartQuantity(prevQuantity => prevQuantity + 1);
+    };
+  
+    const decreaseQuantity = () => {
+      setCartQuantity(prevQuantity => prevQuantity > 1 ? prevQuantity - 1 : 1);
+    };
+
+    const increasePurchaseQuantity = () => {
+        setPurchaseQuantity(prevQuantity => prevQuantity + 1);
+      };
+    
+    const decreasePurchaseQuantity = () => {
+        setPurchaseQuantity(prevQuantity => prevQuantity > 1 ? prevQuantity - 1 : 1);
+    };
+
     useEffect(() => {
         if(id) {
             const url = `/api/books/${id}`;
@@ -33,45 +49,53 @@ function Book() {
     }, [id]);
 
     if(!book) {
-        return <div>Book Not Found!</div>;
+        return <h2>Book Loading...</h2>
     }
 
-    // const deleteBook = async (id) => {
-    //     try {
-    //       await axios.delete(`/api/books/${id}`);
-    //       alert('삭제되었습니다.');
-    //       navigate("/books")
-    //     } catch (error) {
-    //       console.error("삭제에 실패했습니다:", error);
-    //       alert('삭제에 실패했습니다.');
-    //     }
-    // }
-
     const cart = async () => {
-        try {
+        const isConfirmed = window.confirm("장바구니에 " + cartQuantity + "권을 추가하시겠습니까?");
+
+        if(isConfirmed) {
+           try {
             await axios.post(`/api/books/${id}/cart`, {
                 "quantity": cartQuantity,
                 "bookId":id
             });
-            alert('장바구니에 추가되었습니다.');
-        } catch (error) {
-            console.error(error)
-            alert('장바구니 추가에 실패했습니다.');
+            alert("장바구니에 " + cartQuantity  + "권을 추가하였습니다.");
+            window.location.reload();
+            } catch (error) {
+                console.error(error)
+                alert('장바구니 추가에 실패했습니다. 로그인이 필요한 서비스 입니다.');
+                navigate('/login');
+            } 
+        }
+        else {
+            window.location.reload();
         }
     }
 
 
-    const purchase = async () => { // id 파라미터 제거
-        try {
-            await axios.post(`/api/books/${id}/purchase`, {
-                "quantity": purchaseQuantity, // 상태를 사용하여 quantity 설정
-                "bookId": id
-            });
-            alert('구매목록에 추가되었습니다.');
-        } catch (error) {
-            console.error(error);
-            alert('구매목록 추가에 실패했습니다.');
+    const purchase = async () => { 
+        const isConfirmed = window.confirm(purchaseQuantity + "권을 바로 구매하시겠습니까?");
+
+        if(isConfirmed) {
+            try {
+                await axios.post(`/api/books/${id}/purchase`, {
+                    "quantity": purchaseQuantity,
+                    "bookId": id
+                });
+                alert(purchaseQuantity + "권을 구매하였습니다.");
+                window.location.reload();
+            } catch (error) {
+                console.error(error);
+                alert('구매목록 추가에 실패했습니다. 로그인이 필요한 서비스 입니다.');
+                navigate('/login');
+            }  
         }
+        else {
+            window.location.reload();
+        }
+        
     };
 
     const createcomment = async (event) => {
@@ -81,21 +105,33 @@ function Book() {
                 "bookId": id,
                 "content": newcomment
             })
-            alert('댓글이 추가되었습니다.');
+            alert('댓글이 작성되었습니다.');
+            window.location.reload();
         } catch (error) {
             console.error(error)
-            alert('댓글 추가에 실패했습니다.');
+            alert('댓글 작성에 실패했습니다. 로그인이 필요한 서비스 입니다.');
+            navigate('/login');
         }
     }
 
     const deletecomment = async (id) => {
-        try {
+        const isConfirmed = window.confirm("댓글을 삭제 하시겠습니까?");
+
+        if(isConfirmed) {
+            try {
             await axios.delete(`/api/comments/${id}`);
             alert('댓글이 삭제되었습니다.');
+            window.location.reload();
           } catch (error) {
             console.error("삭제에 실패했습니다:", error);
             alert('댓글 삭제에 실패했습니다.');
+            window.location.reload();
           }
+        }
+        else {
+            window.location.reload();
+        }
+        
     }
 
     const startEdit = (comment) => {
@@ -111,11 +147,18 @@ function Book() {
                 "content": editingContent
             });
             alert('댓글이 수정되었습니다.');
+            window.location.reload();
         } catch (error) {
-            console.error("현재 id: "+id + "현재 commentid : " +editingCommentId)
+            console.error("현재 id: "+id + "현재 commentid : " + editingCommentId)
             alert('댓글 수정에 실패했습니다.');
+            window.location.reload();
         }
     };
+
+    const cancelComment = () => {
+        setEditingContent(comment.content);
+        window.location.reload();
+    }
     
     
     return (
@@ -135,7 +178,7 @@ function Book() {
                 <div className="container">
                     <div className="row">
                         <div className="col-md-4" align="center">
-                            <img src={book.imageurl} style={{width:'80%'}}></img>
+                            <img src={book.imageurl} style={{width:'100%'}}></img>
                         </div>
                         
                         <div className="col-md-8" align="center">
@@ -165,14 +208,25 @@ function Book() {
                                     <div className="card-body">
                                         {editingCommentId === comment.id ? (
                                             <>
-                                                <input value={editingContent} onChange={(e) => setEditingContent(e.target.value)} className="form-control"/>
-                                                <button className="btn btn-success" onClick={() => saveComment(comment.id)}>저장</button>
+                                                <div className="d-flex">
+                                                    <div className="col-sm-5">
+                                                        <input value={editingContent} onChange={(e) => setEditingContent(e.target.value)} className="form-control"/>
+                                                    </div>
+                                                    <div className="float-righ">  
+                                                        <button className="btn btn-success" onClick={() => saveComment(comment.id)}>저장</button> 
+                                                        &nbsp;
+                                                        <button className="btn btn-secondary" onClick={cancelComment}>취소</button>
+                                                    </div>
+                                                </div>
+
+                                                {/* <button className="btn btn-success" onClick={() => saveComment(comment.id)}>저장</button> */}
                                             </>
                                         ) : (
                                             <>
                                                 {comment.content}
                                                 <div className="float-right">
                                                     <button className="btn btn-success" onClick={() => startEdit(comment)}>수정 &raquo;</button>              
+                                                    &nbsp;
                                                     <button className="btn btn-danger" onClick={() => deletecomment(comment.id)}>삭제 &raquo;</button>
                                                 </div>
                                             </>
@@ -192,7 +246,7 @@ function Book() {
                             <b>댓글 작성</b>
                             <form onSubmit={createcomment} className="form-horizontal">
                                 <div className="d-flex">
-                                    <div className="col-sm-8">
+                                    <div className="col-sm-5">
                                         <input name="content" value={newcomment} onChange={(e) => setNewcomment(e.target.value)} className="form-control" placeholder="댓글을 입력하세요"/>
                                     </div>
                                     <div className="col-sm-2">  
@@ -201,51 +255,64 @@ function Book() {
                                 </div>                                         
                             </form>
 
-
                             <br />
-                            {/* <Link to={`/books/book/update/${book.id}`} className="btn btn-success" role="button">책 수정 &raquo;</Link>
-                            <Link to={`/books/book/delete/${book.id}`} className="btn btn-danger"  onClick={() => deleteBook(book.id)} role="button">책 삭제 &raquo;</Link> */}
-
                             <br />
                             <br />
 
                             <div>
-                                <label htmlFor="purchaseQuantity">장바구니 수량 선택 : </label>
-                                <select 
-                                    id="purchaseQuantity"
-                                    value={cartQuantity} 
-                                    onChange={e => setCartQuantity(e.target.value)} 
-                                    className="form-control" 
-                                    style={{width: 'auto', display: 'inline-block', margin: '0 10px'}}
-                                >
-                                    {[...Array(10).keys()].map(n => (
-                                        <option key={n + 1} value={n + 1}>{n + 1}</option>
-                                    ))}
-                                </select>
-                                <button className="btn btn-success" onClick={cart}>장바구니에 추가 &raquo;</button>
+                                <label htmlFor="cartQuantity">장바구니 수량 선택 : </label>
+                                <div>
+                                    <button 
+                                        onClick={decreaseQuantity} 
+                                        className="btn btn-secondary" 
+                                        style={{margin: '0 10px'}}
+                                        disabled={cartQuantity === 1} // 1일 때 버튼 비활성화
+                                    >
+                                        -
+                                    </button>
+                                    <span>{cartQuantity}</span>
+                                    <button 
+                                        onClick={increaseQuantity} 
+                                        className="btn btn-secondary" 
+                                        style={{margin: '0 10px'}}
+                                    >
+                                        +
+                                    </button>
+                                    <button className="btn btn-success" onClick={cart}>장바구니에 추가 &raquo;</button>
+                                </div>
+                                
                             </div>
+
+                            <br />
                             
                             <div>
                                 <label htmlFor="purchaseQuantity">구매 수량 선택 : </label>
-                                <select 
-                                    id="purchaseQuantity"
-                                    value={purchaseQuantity} 
-                                    onChange={e => setPurchaseQuantity(e.target.value)} 
-                                    className="form-control" 
-                                    style={{width: 'auto', display: 'inline-block', margin: '0 10px'}}
-                                >
-                                    {[...Array(10).keys()].map(n => (
-                                        <option key={n + 1} value={n + 1}>{n + 1}</option>
-                                    ))}
-                                </select>
-                                <button className="btn btn-danger" onClick={purchase}>바로 구매하기 &raquo;</button>
+                                <div>
+                                    <button 
+                                        onClick={decreasePurchaseQuantity} 
+                                        className="btn btn-secondary" 
+                                        style={{margin: '0 10px'}}
+                                        disabled={purchaseQuantity === 1} // 1일 때 버튼 비활성화
+                                    >
+                                        -
+                                    </button>
+                                    <span>{purchaseQuantity}</span>
+                                    <button 
+                                        onClick={increasePurchaseQuantity} 
+                                        className="btn btn-secondary" 
+                                        style={{margin: '0 10px'}}
+                                    >
+                                        +
+                                    </button>
+                                    <button className="btn btn-danger" onClick={purchase}>바로 구매하기 &raquo;</button>
+                                </div>
                             </div>
                             
                             <br />
                             <br />
 
-                            <Link to={`/books`} className="btn btn-secondary">도서목록 &raquo;</Link>
-                            <Link to={`/cart`} className="btn btn-warning">장바구니 &raquo;</Link>
+                            {/* <Link to={`/books`} className="btn btn-secondary">도서목록 &raquo;</Link>
+                            <Link to={`/cart`} className="btn btn-warning">장바구니 &raquo;</Link> */}
 
                         </div>
                     </div>
