+++ src/pages/order/purchase.js
@@ -2,41 +2,35 @@ import React, { useState, useEffect } from "react";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { useNavigate, useParams, BrowserRouter as Router, Route, Link } from "react-router-dom";
 import axios from 'axios';
+import style from '../cart/style.css';
 
 function Purchase() {
     const [purchase, setPurchase] = useState([]);
     const carturl = `/api/purchase`;
-    // const alldeletecarturl = `/api/books/all/cart`;
-    
+
     useEffect(() => {
         axios.get(`${carturl}`)
         .then(response => setPurchase(response.data))
         .catch(error => console.log(error))
     }, []);
 
-    // const deleteAllCart = async () => {
-    //     try {
-    //         await axios.delete(`${alldeletecarturl}`, {
-    //             "userId":1,
-    //         })
-    //         alert('장바구니에 있는 모든 책들이 삭제되었습니다.');
-    //     } catch (error) {
-    //         console.error(error)
-    //         alert('장바구니에 있는 모든 책들을 삭제하는데 실패했습니다.');
-    //     }
-    // }
-
     const refundbook = async (id) => {
-        try {
-            await axios.delete(`/api/books/${id}/purchase`);
-            alert('해당 책을 구매목록에서 삭제(환불)하였습니다.');
-            window.location.reload();
-        }
-        catch (error) {
-            alert('해당 책을 구매목록에서 삭제(환불)할 수 없습니다.');
-            window.location.reload();
-        }
+        const isConfirmed = window.confirm("해당 책을 환불하시겠습니까?");
 
+        if(isConfirmed) {
+            try {
+                await axios.delete(`/api/books/${id}/purchase`);
+                alert('해당 책을 환불하였습니다.');
+                window.location.reload();
+            }
+            catch (error) {
+                alert('해당 책을 구매목록에서 삭제(환불)할 수 없습니다.');
+            }            
+        }
+        
+        else {
+            window.location.reload();
+        }
     }
 
     return (
@@ -54,33 +48,46 @@ function Purchase() {
                 </div>
 
                 <div className="container">
-                    {purchase && purchase.map((purchase) => (
-                    <div key={purchase.id}>
+                    <div>
+                        <form>
+                            <Link to={`/cart`} className="btn btn-primary float-right" role="button">장바구니 &raquo;</Link> 
+                        </form>
+                        <br />
+                        <br />
+                    </div>
+
+                    <div className="container">
                         <table className="table table-hover">
-                            <tr>
-                                <th>도서ID</th>
-                                <th>도서이름</th>
-                                <th>가격</th>
-                                <th>수량</th>
-                                <th>작가</th>
-                                <th>출판사</th>
-                                <th>도서이미지</th>
-                            </tr>
-                            <tr>
-                                <th>{purchase.bookId}</th>
-                                <th>{purchase.bookname}</th>
-                                <th>{purchase.price}</th>
-                                <th>{purchase.quantity}</th>
-                                <th>{purchase.author}</th>
-                                <th>{purchase.publisher}</th>
-                                <th><img src={purchase.imageUrl} style={{width:'25%'}}></img></th>
-                                <th><button className="btn btn-danger" onClick={() => refundbook(purchase.id)}>삭제</button></th>
-                            </tr>
+                            <thead>
+                                <tr>
+                                    <th>도서이미지</th>
+                                    <th>도서이름</th>
+                                    <th>가격</th>
+                                    <th>수량</th>
+                                    <th>작가</th>
+                                    <th>출판사</th>
+                                    <th>총 가격</th>
+                                    <th>환불</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {purchase && purchase.map((purchase) => (
+                                <tr key={purchase.id}>
+                                    <td><img src={purchase.imageUrl} style={{width:'50%'}} alt="이미지 없음"></img></td>
+                                    <td>{purchase.bookname}</td>
+                                    <td>₩{purchase.price}</td>
+                                    <td>{purchase.quantity}</td>
+                                    <td>{purchase.author}</td>
+                                    <td>{purchase.publisher}</td>
+                                    <td>₩{purchase.price * purchase.quantity}</td>
+                                    <td><button className="btn btn-danger" onClick={() => refundbook(purchase.id)}>환불&raquo;</button></td>
+                                </tr>
+                                ))}   
+                            </tbody>
                         </table>
-                    </div>                        
-                    ))}
+                    </div>
+
                 </div>
-            
             </body>
         </html>
     )
